chore: refactor structure file in project, update readme
