+++ app/react/comparison/page.tsx
@@ -1,4 +1,4 @@
-import HookComparisonsSection from "@/sections/lifecycle/HookComparisonsSection";
+import HookComparisonsSection from "@/sections/react/lifecycle/HookComparisonsSection";
 
 export default function HookComparisonsSectionPage() {
   return <HookComparisonsSection />;
