+++ app/react/lifecycle/page.tsx
@@ -1,4 +1,4 @@
-import LifecycleSection from "@/sections/lifecycle/LifecycleSection";
+import LifecycleSection from "@/sections/react/lifecycle/LifecycleSection";
 
 export default function LifecycleSectionPage() {
   return <LifecycleSection />;
