+++ app/react/virtual-dom/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import VirtualDomSection from "@/sections/virtual-dom/VirtualDomSection";
+import VirtualDomSection from "@/sections/react/virtual-dom/VirtualDomSection";
 
 export default function VirtualDomPage() {
   return <VirtualDomSection />;
